edit maxID to uint32

+++ core/state/candidates/candidates.go
@@ -54,12 +54,12 @@ type RCandidates interface {
 }
 
 type Candidates struct {
-	list map[uint]*Candidate
+	list map[uint32]*Candidate
 
 	isDirty   bool
 	blockList map[types.Pubkey]struct{}
-	pubKeyIDs map[types.Pubkey]uint
-	maxID     uint
+	pubKeyIDs map[types.Pubkey]uint32
+	maxID     uint32
 
 	iavl tree.MTree
 	bus  *bus.Bus
@@ -236,7 +236,7 @@ func (c *Candidates) Create(ownerAddress, rewardAddress, controlAddress types.Ad
 	c.setToMap(pubkey, candidate)
 }
 
-func (c *Candidates) CreateWithID(ownerAddress, rewardAddress, controlAddress types.Address, pubkey types.Pubkey, commission uint, id uint) {
+func (c *Candidates) CreateWithID(ownerAddress, rewardAddress, controlAddress types.Address, pubkey types.Pubkey, commission uint, id uint32) {
 	if id != 0 {
 		c.setPubKeyID(pubkey, id)
 	}
@@ -817,12 +817,12 @@ func (c *Candidates) LoadCandidatesDeliver() {
 
 	_, valueMaxID := c.iavl.Get([]byte{maxIDPrefix})
 	if len(valueMaxID) != 0 {
-		c.maxID = uint(binary.LittleEndian.Uint32(valueMaxID))
+		c.maxID = binary.LittleEndian.Uint32(valueMaxID)
 	}
 
 }
 
-func (c *Candidates) loadCandidatesList() (maxID uint) {
+func (c *Candidates) loadCandidatesList() (maxID uint32) {
 	_, pubIDenc := c.iavl.Get([]byte{pubKeyIDPrefix})
 	if len(pubIDenc) != 0 {
 		var pubIDs []pubkeyID
@@ -830,7 +830,7 @@ func (c *Candidates) loadCandidatesList() (maxID uint) {
 			panic(fmt.Sprintf("failed to decode candidates: %s", err))
 		}
 
-		pubKeyIDs := map[types.Pubkey]uint{}
+		pubKeyIDs := map[types.Pubkey]uint32{}
 		for _, v := range pubIDs {
 			pubKeyIDs[v.PubKey] = v.ID
 			if v.ID > maxID {
@@ -1096,7 +1096,7 @@ func (c *Candidates) setToMap(pubkey types.Pubkey, model *Candidate) {
 	defer c.lock.Unlock()
 
 	if c.list == nil {
-		c.list = map[uint]*Candidate{}
+		c.list = map[uint32]*Candidate{}
 	}
 	c.list[id] = model
 }
@@ -1108,7 +1108,7 @@ func (c *Candidates) setBlockList(blockList map[types.Pubkey]struct{}) {
 	c.blockList = blockList
 }
 
-func (c *Candidates) setPubKeyIDs(list map[types.Pubkey]uint) {
+func (c *Candidates) setPubKeyIDs(list map[types.Pubkey]uint32) {
 	c.lock.Lock()
 	defer c.lock.Unlock()
 
@@ -1200,7 +1200,7 @@ func (c *Candidates) ChangePubKey(old types.Pubkey, new types.Pubkey) {
 	delete(c.pubKeyIDs, old)
 }
 
-func (c *Candidates) getOrNewID(pubKey types.Pubkey) uint {
+func (c *Candidates) getOrNewID(pubKey types.Pubkey) uint32 {
 	c.lock.RLock()
 	id := c.id(pubKey)
 	c.lock.RUnlock()
@@ -1218,23 +1218,23 @@ func (c *Candidates) getOrNewID(pubKey types.Pubkey) uint {
 	return id
 }
 
-func (c *Candidates) id(pubKey types.Pubkey) uint {
+func (c *Candidates) id(pubKey types.Pubkey) uint32 {
 	return c.pubKeyIDs[pubKey]
 }
 
-func (c *Candidates) ID(pubKey types.Pubkey) uint {
+func (c *Candidates) ID(pubKey types.Pubkey) uint32 {
 	c.lock.RLock()
 	defer c.lock.RUnlock()
 
 	return c.pubKeyIDs[pubKey]
 }
 
-func (c *Candidates) setPubKeyID(pubkey types.Pubkey, u uint) {
+func (c *Candidates) setPubKeyID(pubkey types.Pubkey, u uint32) {
 	c.lock.Lock()
 	defer c.lock.Unlock()
 
 	if c.pubKeyIDs == nil {
-		c.pubKeyIDs = map[types.Pubkey]uint{}
+		c.pubKeyIDs = map[types.Pubkey]uint32{}
 	}
 	c.pubKeyIDs[pubkey] = u
 	c.isDirty = true
